AtomBase/python/requirements.txt
langchain-core
langgraph
langchain-ollama
langchain-google-genai
langchain-openai
langchain-anthropic
langchain-huggingface
langchain-groq
python-dotenv
termcolor
tqdm
